+++ sageopt/relaxations/poly_solution_recovery.py
@@ -236,7 +236,6 @@ def _least_squares_magnitude_recovery(con, alpha_reduced, v_reduced, zero_tol):
         constraints.append(tempcon)
     prob = cl.Problem(cl.MIN, t, constraints)
     prob.solve(verbose=False)
-    cl.clear_variable_indices()
     if prob.status in {cl.SOLVED, cl.INACCURATE} and prob.value < np.inf:
         mag = np.exp(y.value.astype(np.longdouble))
         return mag