+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [0.5.4] - unreleased
+## Changed
+ - coniclifts: how ``disjoint_dot`` is computed in ``coniclifts.base``.
+ - interop: GPKit models weren't converted deterministically.
 
 # [0.5.3] - 2020-09-07
 ## Changed